SortManager/src/test/java/com/sparta/jw/sortersTest/BubbleSortTest.java
package com.sparta.jw.sortersTest;

import com.sparta.jw.sorters.BubbleSort;
import com.sparta.jw.sorters.Sortable;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class BubbleSortTest {


    @Test
    void returnsNothingFromEmptyArray(){
        BubbleSort nothing = new BubbleSort();
        int[] input = {};
        nothing.sortArray(input);
    }

    @Test
    void returnsSingleNumberFromSingleNumber(){
        BubbleSort single = new BubbleSort();
        int[] input = {2};
        int[] expected = {2};
        Assertions.assertArrayEquals(expected, single.sortArray(input));
    }

    @Test
    void doesSortThreeTwoOne(){
        BubbleSort threeTwoOne = new BubbleSort();
        int[] input = {3,2,1};
        int[] expected = {1,2,3};
        Assertions.assertArrayEquals(expected, threeTwoOne.sortArray(input));
    }

    @Test
    void doesSortRandom(){
        BubbleSort random = new BubbleSort();
        int[] input = {3,4,23,76,1,4,2,90,23};
        int[] expected = {1,2,3,4,4,23,23,76,90};
        Assertions.assertArrayEquals(expected, random.sortArray(input));
    }

}
